module-5/module-5/requirements.txt
streamlit>=1.29.0
anthropic>=0.25.0
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0 